+++ coregrind/m_sigframe/sigframe-ppc32-linux.c
@@ -752,7 +752,7 @@ void VG_(sigframe_create)( ThreadId tid,
    if (0)
       VG_(printf)("pushed signal frame; %%R1 now = %#lx, "
                   "next %%CIA = %#x, status=%d\n",
-		  sp, tst->arch.vex.guest_CIA, tst->status);
+		  sp, tst->arch.vex.guest_CIA, (Int)tst->status);
 }
 
 
@@ -901,7 +901,7 @@ void VG_(sigframe_destroy)( ThreadId tid, Bool isRT )
 
    if (VG_(clo_trace_signals))
       VG_(message)(Vg_DebugMsg,
-                   "vg_pop_signal_frame (thread %d): "
+                   "vg_pop_signal_frame (thread %u): "
                    "isRT=%d valid magic; EIP=%#x\n",
                    tid, has_siginfo, tst->arch.vex.guest_CIA);
 
@@ -928,7 +928,7 @@ void VG_(sigframe_destroy)( ThreadId tid, Bool isRT )
 //..    if (VG_(clo_trace_signals))
 //..       VG_(message)(
 //..          Vg_DebugMsg, 
-//..          "VG_(signal_return) (thread %d): isRT=%d valid magic; EIP=%p", 
+//..          "VG_(signal_return) (thread %u): isRT=%d valid magic; EIP=%p", 
 //..          tid, isRT, tst->arch.vex.guest_EIP);
 //.. 
 //..    /* tell the tools */
